requirements.txt
pyrogram 
dotenv
google.generativeai
